+++ mne/io/kit/tests/test_kit.py
@@ -10,9 +10,11 @@ import inspect
 import numpy as np
 from numpy.testing import assert_array_almost_equal, assert_array_equal
 from nose.tools import assert_equal, assert_raises, assert_true
+from scipy import linalg
 import scipy.io
 
 from mne import pick_types, Epochs, find_events, read_events
+from mne.transforms import apply_trans
 from mne.tests.common import assert_dig_allclose
 from mne.utils import run_tests_if_main
 from mne.io import Raw, read_raw_kit, read_epochs_kit
@@ -30,8 +32,10 @@ events_path = op.join(data_dir, 'test-eve.txt')
 mrk_path = op.join(data_dir, 'test_mrk.sqd')
 mrk2_path = op.join(data_dir, 'test_mrk_pre.sqd')
 mrk3_path = op.join(data_dir, 'test_mrk_post.sqd')
-elp_path = op.join(data_dir, 'test_elp.txt')
-hsp_path = op.join(data_dir, 'test_hsp.txt')
+elp_txt_path = op.join(data_dir, 'test_elp.txt')
+hsp_txt_path = op.join(data_dir, 'test_hsp.txt')
+elp_path = op.join(data_dir, 'test.elp')
+hsp_path = op.join(data_dir, 'test.hsp')
 
 
 def test_data():
@@ -39,25 +43,25 @@ def test_data():
     """
     assert_raises(TypeError, read_raw_kit, epochs_path)
     assert_raises(TypeError, read_epochs_kit, sqd_path)
-    assert_raises(ValueError, read_raw_kit, sqd_path, mrk_path, elp_path)
+    assert_raises(ValueError, read_raw_kit, sqd_path, mrk_path, elp_txt_path)
     assert_raises(ValueError, read_raw_kit, sqd_path, None, None, None,
                   list(range(200, 190, -1)))
     assert_raises(ValueError, read_raw_kit, sqd_path, None, None, None,
                   list(range(167, 159, -1)), '*', 1, True)
     # check functionality
-    raw_mrk = read_raw_kit(sqd_path, [mrk2_path, mrk3_path], elp_path,
-                           hsp_path)
-    raw_py = _test_raw_reader(read_raw_kit,
-                              input_fname=sqd_path, mrk=mrk_path, elp=elp_path,
-                              hsp=hsp_path, stim=list(range(167, 159, -1)),
-                              slope='+', stimthresh=1)
+    raw_mrk = read_raw_kit(sqd_path, [mrk2_path, mrk3_path], elp_txt_path,
+                           hsp_txt_path)
+    raw_py = _test_raw_reader(read_raw_kit, input_fname=sqd_path, mrk=mrk_path,
+                              elp=elp_txt_path, hsp=hsp_txt_path,
+                              stim=list(range(167, 159, -1)), slope='+',
+                              stimthresh=1)
     assert_true('RawKIT' in repr(raw_py))
     assert_equal(raw_mrk.info['kit_system_id'], KIT.SYSTEM_NYU_2010)
     assert_true(KIT_CONSTANTS[raw_mrk.info['kit_system_id']] is KIT_NY)
 
     # Test stim channel
-    raw_stim = read_raw_kit(sqd_path, mrk_path, elp_path, hsp_path, stim='<',
-                            preload=False)
+    raw_stim = read_raw_kit(sqd_path, mrk_path, elp_txt_path, hsp_txt_path,
+                            stim='<', preload=False)
     for raw in [raw_py, raw_stim, raw_mrk]:
         stim_pick = pick_types(raw.info, meg=False, ref_meg=False,
                                stim=True, exclude='bads')
@@ -133,7 +137,8 @@ def test_raw_events():
 def test_ch_loc():
     """Test raw kit loc
     """
-    raw_py = read_raw_kit(sqd_path, mrk_path, elp_path, hsp_path, stim='<')
+    raw_py = read_raw_kit(sqd_path, mrk_path, elp_txt_path, hsp_txt_path,
+                          stim='<')
     raw_bin = Raw(op.join(data_dir, 'test_bin_raw.fif'))
 
     ch_py = raw_py._raw_extras[0]['sensor_locs'][:, :5]
@@ -151,10 +156,33 @@ def test_ch_loc():
 
     # test when more than one marker file provided
     mrks = [mrk_path, mrk2_path, mrk3_path]
-    read_raw_kit(sqd_path, mrks, elp_path, hsp_path, preload=False)
+    read_raw_kit(sqd_path, mrks, elp_txt_path, hsp_txt_path, preload=False)
     # this dataset does not have the equivalent set of points :(
     raw_bin.info['dig'] = raw_bin.info['dig'][:8]
     raw_py.info['dig'] = raw_py.info['dig'][:8]
     assert_dig_allclose(raw_py.info, raw_bin.info)
+
+
+def test_hsp_elp():
+    """Test KIT usage of *.elp and *.hsp files against *.txt files
+    """
+    raw_txt = read_raw_kit(sqd_path, mrk_path, elp_txt_path, hsp_txt_path)
+    raw_elp = read_raw_kit(sqd_path, mrk_path, elp_path, hsp_path)
+
+    # head points
+    pts_txt = np.array([dig_point['r'] for dig_point in raw_txt.info['dig']])
+    pts_elp = np.array([dig_point['r'] for dig_point in raw_elp.info['dig']])
+    assert_array_almost_equal(pts_elp, pts_txt, decimal=5)
+
+    # transforms
+    trans_txt = raw_txt.info['dev_head_t']['trans']
+    trans_elp = raw_elp.info['dev_head_t']['trans']
+    assert_array_almost_equal(trans_elp, trans_txt, decimal=5)
+
+    # head points in device space
+    pts_txt_in_dev = apply_trans(linalg.inv(trans_txt), pts_txt)
+    pts_elp_in_dev = apply_trans(linalg.inv(trans_elp), pts_elp)
+    assert_array_almost_equal(pts_elp_in_dev, pts_txt_in_dev, decimal=5)
+
 
 run_tests_if_main()